rm react-router, styled-components fix file structure App > /components/App Home > /pages/Home
