+++ client/package.json
@@ -9,10 +9,8 @@
         "@testing-library/user-event": "^13.5.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "react-router": "^6.8.1",
         "react-router-dom": "^6.8.1",
         "react-scripts": "5.0.1",
-        "styled-components": "^5.3.6",
         "typescript": "^4.9.5"
     },
     "devDependencies": {
@@ -20,8 +18,7 @@
         "@types/node": "^18.13.0",
         "@types/react": "^18.0.28",
         "@types/react-dom": "^18.0.10",
-        "@types/react-router": "^5.1.20",
-        "@types/styled-components": "^5.1.26"
+        "@types/react-router-dom": "^5.3.3"
     },
     "scripts": {
         "start": "react-scripts start",
