--- client/src/components/pages/Home/data/countriesData.json
@@ -1,16 +1,0 @@
-{
-    "countriesArray": [
-        {
-            "displayText": "United States",
-            "valueText": "US"
-        },
-        {
-            "displayText": "Canada",
-            "valueText": "CA"
-        },
-        {
-            "displayText": "Mexico",
-            "valueText": "MX"
-        }
-    ]
-}
