+++ client/src/pages/Home/data/countriesData.json
@@ -1,16 +1,16 @@
-{"countriesArray":
-  [
-    {
-      "displayText": "United States",
-      "valueText": "US"
-    },
-    {
-      "displayText": "Canada",
-      "valueText": "CA"
-    },
-    {
-      "displayText": "Mexico",
-      "valueText": "MX"
-    }
-  ]
+{
+    "countriesArray": [
+        {
+            "displayText": "United States",
+            "valueText": "US"
+        },
+        {
+            "displayText": "Canada",
+            "valueText": "CA"
+        },
+        {
+            "displayText": "Mexico",
+            "valueText": "MX"
+        }
+    ]
 }
