--- client/src/pages/Home/data/data.types.tsx
@@ -1,4 +1,0 @@
-export interface filterArrayType {
-  displayText: string;
-  valueText: string;
-}
